+++ Bank/Bank/Bank/Models/Account.cs
@@ -3,9 +3,9 @@
 public class Account
 {
     public Guid Id { get; set; }
-    public required string AccountNumber { get; set; }
+    public required string AccountNumber { get; set; } // Broj žiro računa
     public required string AccountHolderName { get; set; }
-    public decimal Balance { get; set; }
+    public double Balance { get; set; }
 
     public List<DebitCard> DebitCards { get; set; } = new();
